Refactor all the FFI code
